requirements.txt
Flask                 # основной фреймворк для веб-разработки
Flask-SQLAlchemy      # для работы с базой данных SQLite через ORM
Flask-WTF             # для удобной работы с формами
Flask-Login           # для аутентификации и управления сессиями пользователей
Flask-Migrate         # для миграций базы данных
python-dotenv         # для работы с переменными окружения (например, хранение API-ключа OpenAI)
openai                # для работы с API OpenAI и использования ChatGPT
gunicorn              # WSGI-сервер для деплоя Flask-приложения
